Fix bug in categorizing rules_docker-v0.9.0.tar.gz
- Split Categorize method out to a new module
- Add a test for it. The sample set is too large, but it would take a
  long time to cull it for little purpose.
